Add support to custom gitignore files.
This adds two new commands: `joe add` and `joe remove`. This way you can
add your own custom gitignore files to joe.

In case of adding a custom file already available in the default list,
the custom one will be used. This should work fine in Windows too, but I
wasn't able to test it. Any feedback is apreciated.

A new dependency was added: termcolor, for colored output.

The README file still needs to be updated.