Change the indexes for the files
Since now we are using colored displays, we need an index to keep the
filename and the cases since in a Case-Sensitive system this would cause
us trouble. I've tested this in Mac OS X and Linux, the first one case
insensite and the second case sensitive. It's working as expected.